requirements.txt
Django==2.2.1
djangorestframework==3.9.4
django-cors-headers==2.1.0
django-filter==2.1.0
django-celery==3.3.0
django-cors-headers=2.1.0
celery-with-redis==3.0
channels==2.2.0
channels-redis==2.4.0
mysqlclient==1.4.2.post1
redis==3.2.1
celery=3.1.26.post2
requests==2.22.0
kazoo==2.6.1
